ESP32 Board target compilation fixes
The includes for ESP8266 need to be repeated in the ESP32 section.

Also the typedefs u8/u16/u32/u64 are not existing on this platform.
Replace with corresponding standard types uint8_t/uint16_t/uint32_t/uint64_t.

The hardware SHA1 support has a slightly differnt API (plus there
was a typo in the ifdef).