Cleanup temporary string concatenation/creation in debug statements
A lot of heap churn and code overhead is being created by
the debug_* statements that include several temporary String creations
and concatenations. We can clean this up by adding two parameter
overloads that do the concatenation within the function instead
of on the caller side. In order to reduce number of parameters
I created a new naming scheme:

  debug_outln(..., DEBUG_ERROR) -> debug_outln_error(...)
  debug_outln(..., DEBUG_MIN_INFO) -> debug_outln_info(...)
  debug_outln(..., DEBUG_MED_INFO) -> debug_outln_verbose(...)

There is no need for a warning_outln as DEBUG_WARNING is unused
throughout the code. The DEBUG_MAX_INFO lines seem to be not
adding a lot of value (usually they're immediately followed
by a DEBUG_MED_INFO line), so they haven't been converted.
Suggest to remove them instead.

Also remove a couple of unused global variables or variables
only used within one function's context. Although that increased
code size slightly, it overall makes the code easier to follow.

Saves ~ 3500 bytes and almost 100 lines of code in the source.