Merge Adafruit's BMP280/BME280 sensor readings into airrohr-firmware
BMP280 is downwards register compatible to BME280, which features
additionally a humidity sensor. By linking both adafruit libraries
and instanciating them twice we add quite a bit of overhead. Instead
we can differenciate the sensors by the sensor id (which is 0x68 for
BME280 and 0x58 for BMP280) and treat them equivalent.