Switch GPS SoftwareSerial to be only allocated on demand
I would consider the use of GPS rather exotic with a mostly stationary
device, and the SoftwareSerial implementation uses very large buffers
for this. Overall this binds 24kb, or roughly 50% of the total free
RAM. I would consider that fairly excessive.

Most likely the buffer sizes are way too large, and could be reduced
(longest buffer message from GPS seems to be around 78 bytes). Most likely
the large buffer size was considered necessary because the serial was
only read every SAMPLETIME_MS instead of as quickly as possible. Changing
the logic around that