Reduce code duplication in configuration reading
The Password and Char handling can also be done via the configShape
descriptions by adding an additional type "Password" and handling
it accordingly. In order to avoid buffer overflows we're setting
all sizes for password cfg fields to the same variable (it was
that way already anyway). Similarly for other strings we can do the
same, however we still need to fix up the max sizes to avoid
a buffer overflow.

Saves about 1.3kb of Flash.