Add configShape type for Time values
There are two places with clunky code for handling the ms<->seconds
conversation in user interaction. We can simplify that by adding
a generic config type for it. This also allows to reduce the
various helper signatures a bit because more information is
now available in the configShape. This removes ~ 700 bytes of flash
usage.