Switch to Arduino Core 2.6.1
This contains a ton of really nice bugfixes and improvements,
including updates to SoftwareSerial to increase performance
and stability as well as improvements to WiFi handling.

Unfortunately this adds about 10kb of flash size increase, which
we probably can optimize later on down a bit. but we're no
longer constrained by the 512kb limit so that should be just fine.

Memory use goes up by 1%, which is not great but not terrible.