Restore compilation after merge of pull #702
The code that was merged does not compile due to missing references
to "PIN_CS", which have not been defined or documented.

Also the logic for enabling NPM was incorrect, so it always set
serial baud to 115200, while it only wanted to do that for NPM.
so change from defined to defined != 0.

This restores a working firmware, albeit disables NPM support
set NPM_READ to 1 again and please fix it up then.